refactor type strategies to a separate module
